+++ src/RESTPresenterServiceProvider.php
@@ -7,6 +7,8 @@ use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 use XtendPackages\RESTPresenter\Base\RESTPresenter;
 use XtendPackages\RESTPresenter\Commands\Generator\MakeController;
+use XtendPackages\RESTPresenter\Commands\Generator\MakeData;
+use XtendPackages\RESTPresenter\Commands\Generator\MakeFilter;
 use XtendPackages\RESTPresenter\Commands\Generator\MakePresenter;
 use XtendPackages\RESTPresenter\Commands\Generator\MakeResource;
 use XtendPackages\RESTPresenter\Commands\RESTPresenterSetupCommand;
@@ -28,6 +30,8 @@ class RESTPresenterServiceProvider extends PackageServiceProvider
                 MakeResource::class,
                 MakePresenter::class,
                 MakeController::class,
+                MakeFilter::class,
+                MakeData::class,
             ]);
     }